he expected results will be obtaining from both security and performance analyses
The Internet of Things (IoT) has been gaining popularity nowadays due to the increasing usage of smart devices. By relying on the Internet, the IoT could facilitate data sharing between these smart devices. Nevertheless, the involved complex and large-scale network structure in IoT is prone to new security risks. Those conventional access control technologies are not directly applicable in IoT due to their complex access management and lack of credibility owing to centralization. In this research , we put forward a Lattice-based access control mechanism for IoT systems to simplify the access management. For avoiding single point failure and data tampering, the blockchain technology was adopted to distribute the attributes. To attain high efficiency in IoT devices, the access control process was optimized as well. The expected results will be obtaining from both security and performance analyses, which could resist multiple attacks convincingly. Therefore, it is suitable to be adopted in IoT systems.